[RFC] Fix bad rendering on refresh
In #10 I identified some rendering issues on refresh. This is a first
pass at fixing it. It's a bit clumsy, I literally recreate the Document
and replace it, and I think that I might be dropping some CLI options in
the process, but it does at least fix the underlying issue, and now
refreshing no longer has the rendering issues.